+++ ocrd/decorators.py
@@ -2,17 +2,26 @@ import os
 
 import click
 
+from ocrd.constants import VERSION as OCRD_VERSION
 from ocrd.resolver import Resolver
 from ocrd.processor.base import run_processor
 
-def ocrd_cli_wrap_processor(processorClass, mets=None, working_dir=None, cache_enabled=True, *args, **kwargs):
-    if mets.find('://') == -1:
-        mets = 'file://' + mets
-    if mets.startswith('file://') and not os.path.exists(mets[len('file://'):]):
-        raise Exception("File does not exist: %s" % mets)
-    resolver = Resolver(cache_enabled=cache_enabled)
-    workspace = resolver.workspace_from_url(mets, working_dir)
-    run_processor(processorClass, mets, workspace=workspace, *args, **kwargs)
+def ocrd_cli_wrap_processor(processorClass, ocrd_tool=None, mets=None, working_dir=None, cache_enabled=True, dump_json=False, version=False, **kwargs):
+    if dump_json:
+        processorClass(workspace=None, dump_json=True)
+    elif version:
+        p = processorClass(workspace=None)
+        print("Version %s, ocrd/core %s" % (p.version, OCRD_VERSION))
+    elif mets is None:
+        raise Exception('Error: Missing option "-m" / "--mets".')
+    else:
+        if mets.find('://') == -1:
+            mets = 'file://' + mets
+        if mets.startswith('file://') and not os.path.exists(mets[len('file://'):]):
+            raise Exception("File does not exist: %s" % mets)
+        resolver = Resolver(cache_enabled=cache_enabled)
+        workspace = resolver.workspace_from_url(mets, working_dir)
+        run_processor(processorClass, ocrd_tool, mets, workspace=workspace, **kwargs)
 
 def ocrd_cli_options(f):
     """
@@ -28,14 +37,15 @@ def ocrd_cli_options(f):
             print(mets_url)
     """
     params = [
-        click.option('-m', '--mets', help="METS URL to validate", required=True),
+        click.option('-m', '--mets', help="METS URL to validate"),
         click.option('-w', '--working-dir', help="Working Directory"),
         click.option('-I', '--input-file-grp', help='File group(s) used as input.', default='INPUT'),
         click.option('-O', '--output-file-grp', help='File group(s) used as output.', default='OUTPUT'),
         click.option('-g', '--group-id', help="mets:file GROUPID"),
-        click.option('-o', '--output-mets', help="METS URL to write resulting METS to"),
         click.option('-p', '--parameter', type=click.Path()),
+        click.option('-J', '--dump-json', help="Dump tool description as JSON and exit", is_flag=True, default=False),
         click.option('-l', '--log-level', help="Log level", type=click.Choice(['OFF', 'ERROR', 'WARN', 'INFO', 'DEBUG', 'TRACE']), default='INFO'),
+        click.option('-V', '--version', help="Show version", is_flag=True, default=False)
     ]
     for param in params:
         param(f)