+++ ocrd/model/ocrd_mets.py
@@ -1,12 +1,12 @@
-from ocrd.constants import NAMESPACES as NS, TAG_METS_FILE, TAG_METS_FILEGRP, IDENTIFIER_PRIORITY
+from ocrd.constants import NAMESPACES as NS, TAG_METS_FILE, TAG_METS_FILEGRP, IDENTIFIER_PRIORITY, TAG_MODS_IDENTIFIER
 
 from .ocrd_xml_base import OcrdXmlDocument, ET
 from .ocrd_file import OcrdFile
 
 class OcrdMets(OcrdXmlDocument):
 
-    def __init__(self, file_by_id=None, *args, **kwargs):
-        super(OcrdMets, self).__init__(*args, **kwargs)
+    def __init__(self, file_by_id=None, **kwargs):
+        super(OcrdMets, self).__init__(**kwargs)
         if file_by_id is None:
             file_by_id = {}
         self._file_by_id = file_by_id
@@ -21,15 +21,27 @@ class OcrdMets(OcrdXmlDocument):
             if found is not None:
                 return found.text
 
+    @unique_identifier.setter
+    def unique_identifier(self, purl):
+        for t in IDENTIFIER_PRIORITY:
+            id_el = self._tree.getroot().find('.//mods:identifier[@type="%s"]' % t, NS)
+            break
+        if id_el is None:
+            mods = self._tree.getroot().find('.//mods:mods', NS)
+            id_el = ET.SubElement(mods, TAG_MODS_IDENTIFIER)
+            id_el.set('type', 'purl')
+        id_el.text = purl
+
     @property
     def file_groups(self):
         return [el.get('USE') for el in self._tree.getroot().findall('.//mets:fileGrp', NS)]
 
-    def find_files(self, fileGrp=None, groupId=None, mimetype=None):
+    def find_files(self, ID=None, fileGrp=None, groupId=None, mimetype=None):
         """
         List files.
 
         Args:
+            ID (string) : ID of the file
             fileGrp (string) : USE of the fileGrp to list files of
             groupId (string) : GROUPID of matching files
             mimetype (string) : MIMETYPE of matching files
@@ -40,6 +52,8 @@ class OcrdMets(OcrdXmlDocument):
         ret = []
         fileGrp_clause = '' if fileGrp is None else '[@USE="%s"]' % fileGrp
         file_clause = ''
+        if ID is not None:
+            file_clause += '[@ID="%s"]' % ID
         if groupId is not None:
             file_clause += '[@GROUPID="%s"]' % groupId
         if mimetype is not None:
@@ -58,12 +72,16 @@ class OcrdMets(OcrdXmlDocument):
         el_fileGrp.set('USE', fileGrp)
         return el_fileGrp
 
-    def add_file(self, fileGrp, mimetype=None, url=None, ID=None, local_filename=None):
+    def add_file(self, fileGrp, mimetype=None, url=None, ID=None, groupId=None, local_filename=None):
         el_fileGrp = self._tree.getroot().find(".//mets:fileGrp[@USE='%s']" % (fileGrp), NS)
         if el_fileGrp is None:
             el_fileGrp = self.add_file_group(fileGrp)
+        if ID is not None:
+            if self.find_files(ID=ID) != []:
+                raise Exception("File with ID='%s' already exists" % ID)
         mets_file = OcrdFile(ET.SubElement(el_fileGrp, TAG_METS_FILE))
         mets_file.url = url
+        mets_file.groupId = groupId
         mets_file.mimetype = mimetype
         mets_file.ID = ID
         mets_file.local_filename = local_filename