+++ ocrd/processor/base.py
@@ -1,6 +1,8 @@
 import json
 import subprocess
 from ocrd.utils import getLogger
+from ocrd.validator import ParameterValidator
+
 log = getLogger('ocrd.processor')
 
 def _get_workspace(workspace=None, resolver=None, mets_url=None, working_dir=None):
@@ -14,6 +16,7 @@ def _get_workspace(workspace=None, resolver=None, mets_url=None, working_dir=Non
 
 def run_processor(
         processorClass,
+        ocrd_tool=None,
         mets_url=None,
         resolver=None,
         workspace=None,
@@ -21,7 +24,6 @@ def run_processor(
         log_level=None,
         input_file_grp=None,
         output_file_grp=None,
-        output_mets=None,
         parameter=None,
         working_dir=None,
 ):
@@ -31,16 +33,30 @@ def run_processor(
     Args:
         parameter (string): URL to the parameter
     """
-    workspace = _get_workspace(workspace, resolver, mets_url, working_dir)
+    workspace = _get_workspace(
+        workspace,
+        resolver,
+        mets_url,
+        working_dir
+    )
     if parameter is not None:
         fname = workspace.download_url(parameter)
         with open(fname, 'r') as param_json_file:
             parameter = json.load(param_json_file)
+    else:
+        parameter = {}
     log.debug("Running processor %s", processorClass)
-    processor = processorClass(workspace, input_file_grp=input_file_grp, output_file_grp=output_file_grp, parameter=parameter)
+    processor = processorClass(
+        workspace,
+        ocrd_tool=ocrd_tool,
+        group_id=group_id,
+        input_file_grp=input_file_grp,
+        output_file_grp=output_file_grp,
+        parameter=parameter
+    )
     log.debug("Processor instance %s", processor)
     processor.process()
-    #  workspace.persist()
+    workspace.save_mets()
 
 def run_cli(
         binary,
@@ -51,7 +67,6 @@ def run_cli(
         log_level=None,
         input_file_grp=None,
         output_file_grp=None,
-        output_mets=None,
         parameter=None,
         working_dir=None,
 ):
@@ -62,10 +77,9 @@ def run_cli(
     log.debug("Running binary '%s'", binary)
     subprocess.call([
         binary,
-        '-m', mets_url,
-        '-w', workspace.directory
+        '--mets', mets_url,
+        '--working-dir', workspace.directory
     ])
-    #  workspace.persist()
 
 class Processor(object):
     """
@@ -74,11 +88,28 @@ class Processor(object):
     parameter.
     """
 
-    def __init__(self, workspace, parameter=None, input_file_grp="INPUT", output_file_grp="OUTPUT"):
+    def __init__(
+            self,
+            workspace,
+            ocrd_tool=None,
+            parameter={},
+            input_file_grp="INPUT",
+            output_file_grp="OUTPUT",
+            group_id=None,
+            dump_json=False,
+            version=None
+    ):
+        if dump_json:
+            print(json.dumps(ocrd_tool, indent=True))
+            return
+        self.version = version
         self.workspace = workspace
         self.input_file_grp = input_file_grp
         self.output_file_grp = output_file_grp
-        self.parameter = parameter if parameter is not None else {}
+        self.group_id = None if group_id == [] or group_id is None else group_id
+        parameterValidator = ParameterValidator(ocrd_tool)
+        parameterValidator.validate(parameter)
+        self.parameter = parameter
 
     def verify(self):
         """
@@ -97,15 +128,20 @@ class Processor(object):
         """
         List the input files
         """
-        return self.workspace.mets.find_files(fileGrp=self.input_file_grp)
+        return self.workspace.mets.find_files(fileGrp=self.input_file_grp, groupId=self.group_id)
 
-    def add_output_file(self, input_file=None, basename=None, ID=None, **kwargs):
+    def add_output_file(self, basename=None, file_grp=None, ID=None, **kwargs):
         """
         Add an output file.
+
+        Args:
+            basename (string) : basename of the file
+            file_grp (string) : fileGrp to add this file to. Default: self.output_file_grp
+            ID (string) : file@ID
         """
-        log.debug("Adding output file %s", input_file)
-        if basename is None and input_file is not None:
-            basename = input_file.basename_without_extension + '.xml'
-        #  if ID is None and input_file is not None:
-        #      basename = input_file.ID + self.output_file_grp
-        self.workspace.add_file(self.output_file_grp, basename=basename, ID=ID, **kwargs)
+        if basename is None:
+            raise Exception("Must give 'basename' for add_output_file")
+        log.debug("Adding output file %s", basename)
+        if file_grp is None:
+            file_grp = self.output_file_grp
+        self.workspace.add_file(file_grp, basename=basename, ID=ID, **kwargs)