+++ ocrd/utils.py
@@ -1,6 +1,21 @@
+"""
+* xywh_from_points, points_from_xywh, polygon_from_points
+
+The functions have the syntax X_from_Y, where X/Y can be
+
+points a string encoding a polygon: "0,0 100,0 100,100, 0,100"
+polygon an array of x-y-tuples of a polygon: [[0,0], [100,0], [100,100], [0,100]]
+xywh a dict with keys for x, y, width and height: {'x': 0, 'y': 0, 'w': 100, 'h': 100}
+points is for PAGE
+
+polygon is what opencv2 expects
+
+xywh is what tesserocr expects/produces.
+"""
 import subprocess
 import logging
 import re
+import sys
 
 logging.basicConfig(level=logging.DEBUG)
 #  logging.getLogger('ocrd.resolver').setLevel(logging.INFO)
@@ -10,23 +25,44 @@ logging.getLogger('ocrd.resolver.add_files_to_mets').setLevel(logging.INFO)
 def getLogger(*args, **kwargs):
     return logging.getLogger(*args, **kwargs)
 
-def coordinate_string_from_xywh(box):
+def points_from_xywh(box):
     """
     Constructs a polygon representation from a rectangle described as a dict with keys x, y, w, h.
     """
+    x, y, w, h = box['x'], box['y'], box['w'], box['h']
     # tesseract uses a different region representation format
     return "%i,%i %i,%i %i,%i %i,%i" % (
-        box['x'],
-        box['y'],
-        box['x'] + box['w'],
-        box['y'] + box['w'],
-        box['x'] + box['w'] + box['h'],
-        box['y'] + box['w'] + box['h'],
-        box['x'] + box['h'],
-        box['y'] + box['h']
+        x, y,
+        x + w, y,
+        x + w, y + h,
+        x, y + h
     )
 
-def xywh_from_coordinate_string(points):
+def points_from_x0y0x1y1(xyxy):
+    """
+    Constructs a polygon representation from a rectangle described as a list [x0, y0, x1, y1]
+    """
+    [x0, y0, x1, y1] = xyxy
+    return "%s,%s %s,%s %s,%s %s,%s" % (
+        x0, y0,
+        x1, y0,
+        x1, y1,
+        x0, y1
+    )
+
+def xywh_from_points(points):
+    """
+    Constructs an dict representing a rectangle with keys x, y, w, h
+    """
+    [tl, tr, br] = [[int(p) for p in pair.split(',')] for pair in points.split(' ')[:3]]
+    return {
+        'x': tl[0],
+        'y': tl[1],
+        'w': tr[0] - tl[0],
+        'h': br[1] - tr[1],
+    }
+
+def polygon_from_points(points):
     """
     Constructs a numpy-compatible polygon from a page representation.
     """
@@ -46,9 +82,21 @@ def xmllint_format(xml):
     output, _ = proc.communicate(xml)
     return output
 
-def mets_file_id(grp, n):
-    return "%s_%04i"  % (grp, n + 1)
+def is_string(val):
+    return isinstance(val, (str, unicode)) if sys.version_info < (3, 0) else isinstance(val, str)
+
+def concat_padded(base, *args):
+    """
+    Concatenate string and zero-padded 4 digit number
+    """
+    ret = base
+    for n in args:
+        if is_string(n):
+            ret = "%s_%s" % (ret, n)
+        else:
+            ret = "%s_%04i"  % (ret, n + 1)
+    return ret
 
 def safe_filename(url):
-    ret = re.sub('[^A-Za-z0-9]', '', url)
+    ret = re.sub('[^A-Za-z0-9]+', '.', url)
     return ret