+++ test/model/test_ocrd_mets.py
@@ -1,15 +1,23 @@
-from ocrd.model import OcrdMets
-
 from test.base import TestCase, main, assets
-METS_HEROLD = assets.url_of('SBB0000F29300010000/mets.xml')
+
+from ocrd.constants import MIMETYPE_PAGE, METS_XML_EMPTY
+from ocrd.model import OcrdMets
 
 class TestOcrdMets(TestCase):
 
     def setUp(self):
-        self.mets = OcrdMets(filename=METS_HEROLD)
+        self.mets = OcrdMets(filename=assets.url_of('SBB0000F29300010000/mets.xml'))
 
     def test_unique_identifier(self):
         self.assertEqual(self.mets.unique_identifier, 'http://resolver.staatsbibliothek-berlin.de/SBB0000F29300010000', 'Right identifier')
+        self.mets.unique_identifier = 'foo'
+        self.assertEqual(self.mets.unique_identifier, 'foo', 'Right identifier after change')
+
+    def test_unique_identifier_from_nothing(self):
+        self.mets = OcrdMets(content=METS_XML_EMPTY)
+        self.assertEqual(self.mets.unique_identifier, None, 'no identifier')
+        self.mets.unique_identifier = 'foo'
+        self.assertEqual(self.mets.unique_identifier, 'foo', 'Right identifier after change')
 
     def test_file_groups(self):
         self.assertEqual(len(self.mets.file_groups), 17, '17 file groups')
@@ -17,8 +25,8 @@ class TestOcrdMets(TestCase):
     def test_find_files(self):
         self.assertEqual(len(self.mets.find_files(fileGrp='OCR-D-IMG')), 2, '2 files in "OCR-D-IMG"')
         self.assertEqual(len(self.mets.find_files(groupId='FILE_0001_IMAGE')), 17, '17 files with GROUPID "FILE_0001_IMAGE"')
-        self.assertEqual(len(self.mets.find_files(mimetype='image/tif')), 12, '12 image/tif')
-        self.assertEqual(len(self.mets.find_files(mimetype='text/xml')), 22, '22 text/xml')
+        self.assertEqual(len(self.mets.find_files(mimetype='image/tiff')), 12, '12 image/tiff')
+        self.assertEqual(len(self.mets.find_files(mimetype=MIMETYPE_PAGE)), 20, '20 ' + MIMETYPE_PAGE)
         self.assertEqual(len(self.mets.find_files()), 34, '34 files total')
 
     def test_add_group(self):
@@ -29,9 +37,22 @@ class TestOcrdMets(TestCase):
     def test_add_file(self):
         self.assertEqual(len(self.mets.file_groups), 17, '17 file groups')
         self.assertEqual(len(self.mets.find_files(fileGrp='OUTPUT')), 0, '0 files in "OUTPUT"')
-        self.mets.add_file('OUTPUT', mimetype="bla/quux")
+        f = self.mets.add_file('OUTPUT', mimetype="bla/quux", groupId="foobar")
+        self.assertEqual(f.groupId, 'foobar', 'GROUPID set')
         self.assertEqual(len(self.mets.file_groups), 18, '18 file groups')
         self.assertEqual(len(self.mets.find_files(fileGrp='OUTPUT')), 1, '1 files in "OUTPUT"')
+
+    def test_add_file_no_groupid(self):
+        f = self.mets.add_file('OUTPUT', mimetype="bla/quux")
+        self.assertEqual(f.groupId, None, 'No GROUPID')
+
+    def test_add_file_ID_fail(self):
+        f = self.mets.add_file('OUTPUT', ID='best-id-ever')
+        self.assertEqual(f.ID, 'best-id-ever', "ID kept")
+        with self.assertRaises(Exception) as cm:
+            self.mets.add_file('OUTPUT', ID='best-id-ever')
+        self.assertEquals(str(cm.exception), "File with ID='best-id-ever' already exists")
+
 
     def test_file_groupid(self):
         f = self.mets.find_files()[0]