+++ test/model/test_ocrd_page.py
@@ -1,23 +1,68 @@
-from ocrd.model import OcrdPage
-
 from test.base import TestCase, main, assets
-METS_HEROLD_PAGE_5 = assets.url_of('SBB0000F29300010000/00000005.xml')
+import ocrd.model.ocrd_file as ocrd_file
+import ocrd.model.ocrd_page as ocrd_page
+from ocrd.model.ocrd_page import (
+    AlternativeImageType,
+    PcGtsType,
+    PageType,
+    TextRegionType,
+    TextLineType,
+    WordType,
+    GlyphType
+)
 
 # pylint: disable=protected-access
 
 class TestOcrdPage(TestCase):
 
     def setUp(self):
-        self.page = OcrdPage(filename=METS_HEROLD_PAGE_5.replace('file://', ''))
+        with open(assets.path_to('page-with-glyphs.xml'), 'rb') as f:
+            self.xml_as_str = f.read()
+            self.pcgts = ocrd_page.parseString(self.xml_as_str, silence=True)
+
+    def test_from_file(self):
+        f = ocrd_file.OcrdFile(
+            None,
+            mimetype='image/tiff',
+            local_filename=assets.path_to('kant_aufklaerung_1784/kant_aufklaerung_1784_0017.tif')
+        )
+        self.assertEqual(f.mimetype, 'image/tiff')
+        p = ocrd_page.from_file(f)
+        self.assertEqual(p.get_Page().imageWidth, 1457)
 
     def test_pcGtsId(self):
-        self.assertEqual(self.page.pcGtsId, '00000005')
+        self.assertEqual(self.pcgts.pcGtsId, 'glyph-test')
 
     def test_imageFileName(self):
-        self.assertTrue(self.page._tree.find('*[@imageFileName="%s"]'%'foo') is None)
-        self.page.imageFileName = 'foo'
-        self.assertEqual(self.page.imageFileName, 'foo')
-        self.assertFalse(self.page._tree.find('*[@imageFileName="%s"]'%'foo') is None)
+        #  print(self.pcgts.export(sys.stdout, 0))
+        self.assertEqual(self.pcgts.get_Page().imageFilename, '00000259.sw.tif')
+        self.pcgts.get_Page().imageFilename = 'foo'
+        self.assertEqual(self.pcgts.get_Page().imageFilename, 'foo')
+
+    def test_alternativeImage(self):
+        pcgts = PcGtsType(pcGtsId="foo")
+        self.assertEqual(pcgts.pcGtsId, 'foo')
+        # Page/AlternativeImage
+        page = PageType()
+        pcgts.set_Page(page)
+        page.add_AlternativeImage(AlternativeImageType())
+        # TextRegion/AlternativeImage
+        region = TextRegionType()
+        page.add_TextRegion(region)
+        region.add_AlternativeImage(AlternativeImageType())
+        # TextLine/AlternativeImage
+        line = TextLineType()
+        region.add_TextLine(line)
+        line.add_AlternativeImage(AlternativeImageType())
+        # Word/AlternativeImage
+        word = WordType()
+        line.add_Word(word)
+        word.add_AlternativeImage(AlternativeImageType())
+        # Glyph/AlternativeImage
+        glyph = GlyphType()
+        word.add_Glyph(glyph)
+        glyph.add_AlternativeImage(AlternativeImageType())
+
 
 if __name__ == '__main__':
     main()