+++ test/test_utils.py
@@ -1,0 +1,48 @@
+from test.base import TestCase, main
+from ocrd.utils import (
+    points_from_xywh,
+    is_string,
+    concat_padded,
+    points_from_x0y0x1y1,
+    xywh_from_points,
+    polygon_from_points
+)
+
+class TestResolver(TestCase):
+
+    #  def runTest(self):
+
+    def test_points_from_xywh(self):
+        self.assertEqual(
+            points_from_xywh({'x': 100, 'y': 100, 'w': 100, 'h': 100}),
+            '100,100 200,100 200,200 100,200'
+        )
+
+    def test_points_from_x0y0x1y1(self):
+        self.assertEqual(
+            points_from_x0y0x1y1([100, 100, 200, 200]),
+            '100,100 200,100 200,200 100,200'
+        )
+
+
+    def test_xywh_from_points(self):
+        self.assertEqual(
+            xywh_from_points('100,100 200,100 200,200 100,200'),
+            {'x': 100, 'y': 100, 'w': 100, 'h': 100})
+
+    def test_polygon_from_points(self):
+        self.assertEqual(
+            polygon_from_points('100,100 200,100 200,200 100,200'),
+            [[100, 100], [200, 100], [200, 200], [100, 200]])
+
+    def test_concat_padded(self):
+        self.assertEqual(concat_padded('x', 0), 'x_0001')
+        self.assertEqual(concat_padded('x', 0, 1, 2), 'x_0001_0002_0003')
+        self.assertEqual(concat_padded('x', 0, '1', 2), 'x_0001_1_0003')
+
+    def test_is_string(self):
+        self.assertTrue(is_string('x'))
+        self.assertTrue(is_string(u'x'))
+
+if __name__ == '__main__':
+    main()